# docs/openapi.yml
openapi: 3.0.3
info:
  title: Deck Tracker API
  version: 1.0.0
  description: Track deck building runs across tabletop and roguelike games

paths:

  /games:
    get:
      summary: List all games
      operationId: getGames
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameSummary'

    post:
      summary: Add a game
      operationId: createGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSummary'
        '400':
          $ref: '#/components/responses/BadRequest'

  /games/{gameId}:
    get:
      summary: Get a game with its card pool
      operationId: getGame
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Replace a game
      operationId: updateGame
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a game and all associated data
      operationId: deleteGame
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '204':
          description: No content
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /games/{gameId}/card-types:
    get:
      summary: List card types for a game
      operationId: getCardTypes
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardType'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      summary: Add a card type to a game
      operationId: createCardType
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardTypeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /card-types/{cardTypeId}:
    get:
      summary: Get a single card type
      operationId: getCardType
      parameters:
        - $ref: '#/components/parameters/CardTypeId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardType'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Replace a card type
      operationId: updateCardType
      parameters:
        - $ref: '#/components/parameters/CardTypeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardTypeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a card type
      operationId: deleteCardType
      parameters:
        - $ref: '#/components/parameters/CardTypeId'
      responses:
        '204':
          description: No content
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /games/{gameId}/cards:
    get:
      summary: List cards in a game's card pool
      operationId: getCards
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      summary: Add a card to a game's card pool
      operationId: createCard
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /cards/{cardId}:
    get:
      summary: Get a single card
      operationId: getCard
      parameters:
        - $ref: '#/components/parameters/CardId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Replace a card
      operationId: updateCard
      parameters:
        - $ref: '#/components/parameters/CardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a card
      operationId: deleteCard
      parameters:
        - $ref: '#/components/parameters/CardId'
      responses:
        '204':
          description: No content
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /games/{gameId}/analysis:
    get:
      summary: Get aggregate analysis across all decks for a game
      operationId: getGameAnalysis
      parameters:
        - $ref: '#/components/parameters/GameId'
        - name: result
          in: query
          required: false
          description: Filter to only winning or losing decks
          schema:
            $ref: '#/components/schemas/DeckResult'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregateAnalysis'
        '404':
          $ref: '#/components/responses/NotFound'


  /games/{gameId}/decks:
    get:
      summary: List decks, optionally filtered by game
      operationId: getDecks
      parameters:
        - name: result
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DeckResult'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedDecks'

    post:
      summary: Log a new deck
      operationId: createDeck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckDetail'
        '400':
          $ref: '#/components/responses/BadRequest'

  /decks/{deckId}:
    get:
      summary: Get a deck with its cards
      operationId: getDeck
      parameters:
        - $ref: '#/components/parameters/DeckId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Replace a deck including its card list
      operationId: updateDeck
      parameters:
        - $ref: '#/components/parameters/DeckId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a deck
      operationId: deleteDeck
      parameters:
        - $ref: '#/components/parameters/DeckId'
      responses:
        '204':
          description: No content
        '404':
          $ref: '#/components/responses/NotFound'

  /decks/{deckId}/analysis:
    get:
      summary: Get analysis for a single deck
      operationId: getDeckAnalysis
      parameters:
        - $ref: '#/components/parameters/DeckId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckAnalysis'
        '404':
          $ref: '#/components/responses/NotFound'

components:

  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: integer
    CardTypeId:
      name: cardTypeId
      in: path
      required: true
      schema:
        type: integer
    CardId:
      name: cardId
      in: path
      required: true
      schema:
        type: integer
    DeckId:
      name: deckId
      in: path
      required: true
      schema:
        type: integer

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Resource cannot be deleted due to existing references
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

  schemas:

    GameSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        totalDecks:
          type: integer
        winRate:
          type: number
          format: float
          description: 0.0 to 1.0

    GameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 500

    GameDetail:
      allOf:
        - $ref: '#/components/schemas/GameSummary'
        - type: object
          properties:
            cardTypes:
              type: array
              items:
                $ref: '#/components/schemas/CardType'
            cards:
              type: array
              items:
                $ref: '#/components/schemas/Card'

    CardType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true

    CardTypeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 500

    Card:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CardType'
        cost:
          type: integer
          nullable: true
          description: Energy/coin cost — nullable since not all games have this

    CardRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 500
        typeId:
          type: integer
          nullable: true
          description: Must reference a CardType belonging to this game
        cost:
          type: integer
          nullable: true
          minimum: 0

    DeckResult:
      type: string
      enum: [win, loss, abandoned]

    DeckCardEntry:
      type: object
      required: [cardId, quantity]
      properties:
        cardId:
          type: integer
        quantity:
          type: integer
          minimum: 1

    DeckSummary:
      type: object
      properties:
        id:
          type: integer
        gameId:
          type: integer
        gameName:
          type: string
        result:
          $ref: '#/components/schemas/DeckResult'
        character:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        cardCount:
          type: integer

    DeckRequest:
      type: object
      required: [result, cards]
      properties:
        result:
          $ref: '#/components/schemas/DeckResult'
        character:
          type: string
          nullable: true
          description: Optional — e.g. Ironclad, Silent, Watcher in StS
        notes:
          type: string
          nullable: true
        cards:
          type: array
          description: The final deck — list of card IDs with quantities
          items:
            $ref: '#/components/schemas/DeckCardEntry'

    DeckDetail:
      allOf:
        - $ref: '#/components/schemas/DeckSummary'
        - type: object
          properties:
            notes:
              type: string
              nullable: true
            cards:
              type: array
              items:
                $ref: '#/components/schemas/DeckCardEntry'

    PagedDecks:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DeckSummary'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    CardFrequency:
      type: object
      properties:
        card:
          $ref: '#/components/schemas/Card'
        count:
          type: integer
        winRate:
          type: number
          format: float

    CardPair:
      type: object
      description: Two cards that appear together, with frequency and win rate
      properties:
        cardA:
          $ref: '#/components/schemas/Card'
        cardB:
          $ref: '#/components/schemas/Card'
        coOccurrenceCount:
          type: integer
        winRate:
          type: number
          format: float

    DeckAnalysis:
      type: object
      description: Analysis of a single deck
      properties:
        deckId:
          type: integer
        totalCardCount:
          type: integer
          description: Sum of all card quantities
        uniqueCardCount:
          type: integer
          description: Number of distinct cards
        typeBreakdown:
          type: object
          additionalProperties:
            type: integer
          description: e.g. "Attack"=12, "Skill"=8, "Power"=3
        costCurve:
          type: object
          additionalProperties:
            type: integer
          description: Card count at each cost value e.g. "0"=3, "1"=8, "2"=6, "3+"=4
          nullable: true
          x-note: null if the game has no card costs
        averageCost:
          type: number
          format: float
          nullable: true
        mostDuplicated:
          nullable: true
          description: The card with the highest quantity in this deck
          allOf:
            - $ref: '#/components/schemas/DeckCardEntry'

    AggregateAnalysis:
      type: object
      description: Cross-deck analysis for a game
      properties:
        gameId:
          type: integer
        totalDecks:
          type: integer
        winRate:
          type: number
          format: float
        topCards:
          type: array
          description: Most frequently included cards, sorted by win rate
          items:
            $ref: '#/components/schemas/CardFrequency'
        topPairs:
          type: array
          description: Most common card pairings, sorted by co-occurrence count
          items:
            $ref: '#/components/schemas/CardPair'

    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
        detail:
          type: string